refactor(core): move parse-result to api-elements
