+++ docs/conf.json
@@ -6,7 +6,7 @@
       "../node_modules/minim/lib/elements",
       "../node_modules/minim/lib/serialisers",
       "../packages/minim-api-description/lib/elements",
-      "../packages/minim-parse-result/lib"
+      "../packages/api-elements/lib"
     ]
   }
 }
