--- packages/minim-parse-result/.npmignore
@@ -1,1 +1,0 @@
-test
